# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NHS (Leigh) Member Hour Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export interface MonthlyHours {
  month: string;
  shortLabel: string;
  hwCenter: string;
  tutorial: string;
  total: number;
  hasData: boolean;
}

export interface DashboardData {
  displayName: string;
  grade: number;
  infoFormComplete: boolean;
  clubDuesPaid: boolean;
  totalHours: number;
  annualGoal: number;
  annualRemaining: number;
  semester1Hours: number;
  semester1Goal: number;
  semester1Remaining: number;
  semester2Hours: number;
  semester2Goal: number;
  semester2Remaining: number;
  monthlyHours: MonthlyHours[];
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}