# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NHS (Leigh) Member Hour Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: dashboard
    description: Dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get the current member's dashboard data
      responses:
        "200":
          description: Dashboard data from Google Sheets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found in spreadsheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1

    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string

    DashboardData:
      type: object
      required:
        - displayName
        - grade
        - infoFormComplete
        - clubDuesPaid
        - totalHours
        - annualGoal
        - annualRemaining
        - semester1Hours
        - semester1Goal
        - semester1Remaining
        - semester2Hours
        - semester2Goal
        - semester2Remaining
        - monthlyHours
      properties:
        displayName:
          type: string
        grade:
          type: integer
        infoFormComplete:
          type: boolean
        clubDuesPaid:
          type: boolean
        totalHours:
          type: number
        annualGoal:
          type: number
        annualRemaining:
          type: number
        semester1Hours:
          type: number
        semester1Goal:
          type: number
        semester1Remaining:
          type: number
        semester2Hours:
          type: number
        semester2Goal:
          type: number
        semester2Remaining:
          type: number
        monthlyHours:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyHours"

    MonthlyHours:
      type: object
      required: [month, shortLabel, hwCenter, tutorial, total, hasData]
      properties:
        month:
          type: string
        shortLabel:
          type: string
        hwCenter:
          type: string
        tutorial:
          type: string
        total:
          type: number
        hasData:
          type: boolean

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string